requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis[hiredis]==5.0.1
openai==1.12.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
pydantic==2.6.0